requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
fire==0.7.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
idna==3.11
joblib==1.5.3
mysqlclient==2.2.7
numpy==1.26.4
packaging==25.0
pandas==2.3.3
platformdirs==4.5.1
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.48.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
starlette==0.50.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wandb==0.23.1
